+++ proyectoFinal/proyectoFinal/ViewM/PedidoVM.cs
@@ -26,9 +26,9 @@ namespace proyectoFinal.ViewM
 
         public int cantidad { get; set; }
 
-        public decimal precioUnitario { get; set; }
+        public int precioUnitario { get; set; }
 
-        public decimal subtotal => cantidad * precioUnitario;
+        public int subtotal => cantidad * precioUnitario;
     }
 }
